add more services

+++ src/main/java/com/gouzal/iquote/model/Author.java
@@ -1,6 +1,5 @@
 package com.gouzal.iquote.model;
 
-import com.fasterxml.jackson.annotation.JsonIgnore;
 import lombok.Data;
 import lombok.EqualsAndHashCode;
 import lombok.NoArgsConstructor;
@@ -15,7 +14,7 @@ import java.util.List;
 public class Author {
     @Id
     @GeneratedValue(strategy = GenerationType.AUTO)
-    private int id;
+    private Long id;
     private String fullName;
     @OneToMany(mappedBy = "author")
     List<Quote> quotes;
