+++ src/main/java/com/gouzal/iquote/model/Tag.java
@@ -9,7 +9,6 @@ import lombok.NoArgsConstructor;
 import javax.persistence.*;
 import java.util.List;
 import java.util.Objects;
-import java.util.Set;
 
 @NoArgsConstructor
 @Entity
@@ -19,7 +18,7 @@ import java.util.Set;
 public class Tag {
     @Id
     @GeneratedValue(strategy = GenerationType.AUTO)
-    private int id;
+    private Long id;
     private String value;
     @JsonIgnore
     @ManyToMany(mappedBy = "tags")
