+++ src/main/java/com/gouzal/iquote/repository/QuoteRepository.java
@@ -3,5 +3,5 @@ package com.gouzal.iquote.repository;
 import com.gouzal.iquote.model.Quote;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface QuoteRepository extends JpaRepository<Quote, Integer> {
+public interface QuoteRepository extends JpaRepository<Quote, Long> {
 }
